Move pagination plugin tests to other test files
