+++ test/integration.js
@@ -83,7 +83,6 @@ module.exports = function(Bookshelf) {
       require('./integration/plugins/virtuals')(bookshelf);
       require('./integration/plugins/visibility')(bookshelf);
       require('./integration/plugins/registry')(bookshelf);
-      require('./integration/plugins/pagination')(bookshelf);
       require('./integration/plugins/case-converter')(bookshelf);
       require('./integration/plugins/processor')(bookshelf);
     });
